Put play button inside Game Progress Component
