+++ react/gamearea.js
@@ -76,12 +76,6 @@ class Gamearea extends Component {
       team_id_2: json.team_2_id, 
       display_name: json.display_name,
       startTime: json.start_time });
-    /* console.log(this.state.startTime);
-    var now = moment.utc(Date.now());
-    console.log(now);
-    let countdown = moment.duration(now.diff(this.state.startTime));
-    console.log(countdown._milliseconds);
-    this.setState({countdown: countdown._milliseconds}); */
     this.intervalID = setTimeout(this.updateGame.bind(this), 1000);
   }
 
@@ -90,7 +84,7 @@ class Gamearea extends Component {
       return (
         <div>
           Game Area
-          <Gameprogress startTime={this.state.startTime}  t1Name={ this.state.teamName_1 } t1Begin={4} t1End={9} t2Name={ this.state.teamName_2 } t2Begin={1} t2End={2} />
+          <Gameprogress startTime={this.state.startTime}  t1Name={ this.state.teamName_1 } t1Begin={4} t1End={9} seatsFull={this.state.seatsFull} t2Name={ this.state.teamName_2 } t2Begin={1} t2End={2} />
           <ChairsCollection seats={this.state.seats} gameID={this.props.match.params.gameID} setSeatId={(id) => this.setSeatId(id)} playerSeatId={this.state.seatId} />
         </div>
       );
@@ -99,7 +93,7 @@ class Gamearea extends Component {
         <div>
           Game Area
           <EnvelopeStack></EnvelopeStack>
-          <Gameprogress startTime={this.state.startTime} t1Name={ this.state.teamName_1 } t1Begin={this.state.startTime} t1End={9} t2Name={ this.state.teamName_2 } t2Begin={this.state.startTime} t2End={2} />
+          <Gameprogress facilitatorGets={this.props.location.state.facilitatorID} gameID={this.props.match.params.gameID} startTime={this.state.startTime} t1Name={ this.state.teamName_1 } t1Begin={this.state.startTime} t1End={9} seatsFull={this.state.seatsFull} t2Name={ this.state.teamName_2 } t2Begin={this.state.startTime} t2End={2} />
           <ChairsCollection seats={this.state.seats} gameID={this.props.match.params.gameID} setSeatId={(id) => this.setSeatId(id)} playerSeatId={this.state.seatId} />
           <Controls facilitatorGets={this.props.location.state.facilitatorID} team_id_1={this.state.team_id_1} team_id_2={this.state.team_id_2} seatsFull={this.state.seatsFull} gameID={this.props.match.params.gameID}/>
         </div>
