+++ react/gameprogress.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react';
 import moment from 'moment';
 import Card from 'react-bootstrap/Card';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faPlay, faPause } from '@fortawesome/free-solid-svg-icons'
 
-function doMath(end, begin) {
-  return end - begin;
-}
 
 
 
@@ -12,28 +11,34 @@ class Gameprogress extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      countdown: []
+      isGoing: false
     }
+    this.togglePlay = this.togglePlay.bind(this);
   }
 
 doTime() {
   if (this.props.startTime)  {
-    var now = moment.utc(Date.now());
+    let now = moment.utc(Date.now());
     let countdown = moment.duration(now.diff(this.props.startTime));
-    console.log(countdown);
-    return (<p>{countdown._data.minutes}:{countdown._data.seconds}:{countdown._data.milliseconds}</p>);
+    //return countdown._data.seconds;
+    return (<p>{countdown._data.minutes}:{countdown._data.seconds}</p>);
   }
   else {
     return '';
   }
 }
+
+async togglePlay(val) {
+  this.setState({ isGoing: !this.state.isGoing });
+  const response = await fetch(`/api/start-game/${this.props.facilitatorGets}/${this.props.gameID}`)
+  const json = await response.json();
+  console.log(json);
+}
   
   render() {
-    const t1TimeSpent = doMath(this.props.t1End, this.props.t1Begin);
-    const t2TimeSpent = doMath(this.props.t2End, this.props.t2Begin);
-    //console
-    //var countdown = this.doTime(this.props.startTime);
+    let ic = this.state.isGoing ? faPause : faPlay;
     const profitPerEnvelope = 22;
+    const facilID = this.props.facilitatorGets;
     return (
       <div >
         
@@ -41,13 +46,16 @@ doTime() {
           <Card.Body>
           
             <h1>Money Earned</h1>
+            {facilID && 
+              <FontAwesomeIcon icon={ic} spin onClick={this.props.seatsFull ? this.togglePlay : 'null'}/>
+            }
             <p> {this.props.t1Name} --- VS --- {this.props.t2Name}</p>
             <h2 >
-              ${t1TimeSpent * profitPerEnvelope}
+              ${profitPerEnvelope * 0}
                     &nbsp;&nbsp;&nbsp;&nbsp;
                       {this.doTime()}
                     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                    ${t2TimeSpent * profitPerEnvelope}
+                    ${profitPerEnvelope * 0}
             </h2>
             
           </Card.Body>
